Avoid double config validation
